Introduced HollowIncrementalCyclePopulator and cleanup mutations at the end
